Fixed all mistakes
